Modularized network information scripts
